refactor: Moved general functions to utils
